Remove debug codes & Improve code quality
